component renaming and prettier formatting
