+++ package.json
@@ -7,6 +7,7 @@
     "dev": "vite",
     "build": "tsc && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "format": "prettier . --write",
     "preview": "vite preview"
   },
   "dependencies": {
@@ -22,6 +23,7 @@
     "eslint": "^8.53.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.4",
+    "prettier": "3.1.0",
     "typescript": "^5.2.2",
     "vite": "^5.0.0"
   }
