+++ src/App.tsx
@@ -1,20 +1,29 @@
-import './App.css'
+import "./App.css";
 import profilePicture from "./assets/profile.jpg";
-import { Graphic } from "./Graphic/Graphic";
+import { ParticleCloud } from "./particle_cloud/particle_cloud";
 
-function App() {
+export function App() {
   return (
     <>
       <div className="landing">
-        <Graphic />
+        <ParticleCloud />
         <div className="info">
-          <img src={profilePicture} className="Profile-Pic" alt={"Profile Picture"} />
+          <img
+            src={profilePicture}
+            className="Profile-Pic"
+            alt={"Profile Picture"}
+          />
           <h1>Thomas Northall-Little</h1>
           <p className="byline">Senior Software Engineer at Canva</p>
 
-          <p> <a href="mailto:contact@tomnlittle.com">contact@tomnlittle.com</a></p>
+          <p>
+            {" "}
+            <a href="mailto:contact@tomnlittle.com">contact@tomnlittle.com</a>
+          </p>
           <p className="Links">
-            <a href="https://www.linkedin.com/in/thomas-northall-little-073604111">LinkedIn</a>
+            <a href="https://www.linkedin.com/in/thomas-northall-little-073604111">
+              LinkedIn
+            </a>
             <a href="https://github.com/tomnlittle">GitHub</a>
           </p>
         </div>
@@ -23,7 +32,5 @@ function App() {
         <p> Copyright 2023 - Thomas Northall-Little </p>
       </div>
     </>
-  )
+  );
 }
-
-export default App
