--- src/Graphic/Graphic.css
@@ -1,8 +1,0 @@
-.particle {
-    position: absolute;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    margin: 0;
-    z-index: -1;
-}
