+++ src/index.css
@@ -3,8 +3,8 @@ html {
 }
 
 :root {
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
+    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
     sans-serif;
 
   font-synthesis: none;
@@ -30,11 +30,13 @@ h1 {
   margin-bottom: 4px;
 }
 
-a, a:visited {
+a,
+a:visited {
   text-decoration: none;
   color: #282c34;
 }
 
-a:hover, a:visited:hover {
-  color: #B9009A;
+a:hover,
+a:visited:hover {
+  color: #b9009a;
 }
