+++ src/particle_cloud/particle_cloud.css
@@ -1,0 +1,8 @@
+.particleCloud {
+  position: absolute;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  margin: 0;
+  z-index: -1;
+}
